Refactor ThumbnailWidget
